# contracts/openapi/orchestrator-v0.1.json
{
  "components": {
    "schemas": {
...
      "ContactLookup": {
        "additionalProperties": false,
        "properties": {
          "contact_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Contact Id",
            "type": "string"
          }
        },
        "required": [
          "contact_id"
        ],
        "title": "ContactLookup",
        "type": "object"
      },
      "ContactMetadata": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "criticality": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Criticality"
          },
          "end_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "ephemeris_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ephemeris Id"
          },
          "ephemeris_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ephemeris Mode"
          },
          "external_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Ref"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "mission_profile_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mission Profile Id"
          },
          "properties_cfes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Properties Cfes"
          },
          "properties_is_internal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Properties Is Internal"
          },
          "properties_is_test": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Properties Is Test"
          },
          "reservation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reservation Id"
          },
          "setup_duration": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Setup Duration"
          },
          "signature_comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signature Comment"
          },
          "signature_contains_human_override": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signature Contains Human Override"
          },
          "signature_last_impacting_principal_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signature Last Impacting Principal Kind"
          },
          "signature_outcome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signature Outcome"
          },
          "spacecraft_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spacecraft Id"
          },
          "start_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "station_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Station Id"
          },
          "system_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "System Id"
          },
          "teardown_duration": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Teardown Duration"
          },
          "tenant_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Id"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "title": "ContactMetadata",
        "type": "object"
      },
...
      "DatasetFetchRequest": {
        "additionalProperties": false,
        "properties": {
          "nominal_carrier_frequency_hz": {
            "exclusiveMinimum": 0.0,
            "title": "Nominal Carrier Frequency Hz",
            "type": "number"
          },
          "query": {
            "$ref": "#/components/schemas/DatasetQuery"
          }
        },
        "required": [
          "nominal_carrier_frequency_hz",
          "query"
        ],
        "title": "DatasetFetchRequest",
        "type": "object"
      },
      "DatasetPacket": {
        "additionalProperties": false,
        "properties": {
          "measurements": {
            "items": {
              "$ref": "#/components/schemas/Measurement"
            },
            "title": "Measurements",
            "type": "array"
          },
          "presented_count": {
            "minimum": 0,
            "title": "Presented Count",
            "type": "integer"
          },
          "provenance": {
            "additionalProperties": {
              "type": "string"
            },
            "default": {},
            "title": "Provenance",
            "type": "object"
          },
          "query": {
            "$ref": "#/components/schemas/DatasetQuery"
          },
          "raw_count": {
            "minimum": 0,
            "title": "Raw Count",
            "type": "integer"
          },
          "rejected_count": {
            "minimum": 0,
            "title": "Rejected Count",
            "type": "integer"
          },
          "schema_version": {
            "const": "0.1",
            "default": "0.1",
            "title": "Schema Version",
            "type": "string"
          },
          "stations": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Cartesian3"
            },
            "title": "Stations",
            "type": "object"
          },
          "tdm": {
            "$ref": "#/components/schemas/TdmDocument"
          }
        },
        "required": [
          "tdm",
          "measurements",
          "stations",
          "query",
          "raw_count",
          "presented_count",
          "rejected_count"
        ],
        "title": "DatasetPacket",
        "type": "object"
      },
      "DatasetQuery": {
        "additionalProperties": false,
        "properties": {
          "contact_ids": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Contact Ids",
            "type": "array"
          },
          "end_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "End Time"
          },
          "maximum_doppler_hz": {
            "default": 100000.0,
            "title": "Maximum Doppler Hz",
            "type": "number"
          },
          "minimum_doppler_hz": {
            "default": -100000.0,
            "title": "Minimum Doppler Hz",
            "type": "number"
          },
          "minimum_ebn0_db": {
            "default": 0.0,
            "title": "Minimum Ebn0 Db",
            "type": "number"
          },
          "minimum_elevation_deg": {
            "default": 1.0,
            "title": "Minimum Elevation Deg",
            "type": "number"
          },
          "require_lock": {
            "default": false,
            "title": "Require Lock",
            "type": "boolean"
          },
          "spacecraft_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Spacecraft Id"
          },
          "start_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Start Time"
          }
        },
        "title": "DatasetQuery",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "GatewayBearer": {
        "bearerFormat": "opaque",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "DART Orchestrator and API Proxy",
    "version": "0.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v0/datasets/query": {
      "post": {
        "description": "Fetch a bounded dataset after structural and semantic contract validation.",
        "operationId": "query_dataset_v0_datasets_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasetFetchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetPacket"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "detail": {
                      "maxLength": 2000,
                      "minLength": 1,
                      "type": "string"
                    },
                    "instance": {
                      "maxLength": 2000,
                      "type": "string"
                    },
                    "status": {
                      "maximum": 599.0,
                      "minimum": 400.0,
                      "type": "integer"
                    },
                    "title": {
                      "maxLength": 200,
                      "minLength": 1,
                      "type": "string"
                    },
                    "type": {
                      "enum": [
                        "urn:dart:problem:authentication",
                        "urn:dart:problem:configuration",
                        "urn:dart:problem:conflict",
                        "urn:dart:problem:domain",
                        "urn:dart:problem:internal",
                        "urn:dart:problem:method-not-allowed",
                        "urn:dart:problem:not-found",
                        "urn:dart:problem:upstream",
                        "urn:dart:problem:validation"
                      ],
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Bearer authentication failed."
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "detail": {
                      "maxLength": 2000,
                      "minLength": 1,
                      "type": "string"
                    },
                    "instance": {
                      "maxLength": 2000,
                      "type": "string"
                    },
                    "status": {
                      "maximum": 599.0,
                      "minimum": 400.0,
                      "type": "integer"
                    },
                    "title": {
                      "maxLength": 200,
                      "minLength": 1,
                      "type": "string"
                    },
                    "type": {
                      "enum": [
                        "urn:dart:problem:authentication",
                        "urn:dart:problem:configuration",
                        "urn:dart:problem:conflict",
                        "urn:dart:problem:domain",
                        "urn:dart:problem:internal",
                        "urn:dart:problem:method-not-allowed",
                        "urn:dart:problem:not-found",
                        "urn:dart:problem:upstream",
                        "urn:dart:problem:validation"
                      ],
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "The request is syntactically or semantically invalid."
          },
          "502": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "detail": {
                      "maxLength": 2000,
                      "minLength": 1,
                      "type": "string"
                    },
                    "instance": {
                      "maxLength": 2000,
                      "type": "string"
                    },
                    "status": {
                      "maximum": 599.0,
                      "minimum": 400.0,
                      "type": "integer"
                    },
                    "title": {
                      "maxLength": 200,
                      "minLength": 1,
                      "type": "string"
                    },
                    "type": {
                      "enum": [
                        "urn:dart:problem:authentication",
                        "urn:dart:problem:configuration",
                        "urn:dart:problem:conflict",
                        "urn:dart:problem:domain",
                        "urn:dart:problem:internal",
                        "urn:dart:problem:method-not-allowed",
                        "urn:dart:problem:not-found",
                        "urn:dart:problem:upstream",
                        "urn:dart:problem:validation"
                      ],
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "An upstream dependency failed."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "detail": {
                      "maxLength": 2000,
                      "minLength": 1,
                      "type": "string"
                    },
                    "instance": {
                      "maxLength": 2000,
                      "type": "string"
                    },
                    "status": {
                      "maximum": 599.0,
                      "minimum": 400.0,
                      "type": "integer"
                    },
                    "title": {
                      "maxLength": 200,
                      "minLength": 1,
                      "type": "string"
                    },
                    "type": {
                      "enum": [
                        "urn:dart:problem:authentication",
                        "urn:dart:problem:configuration",
                        "urn:dart:problem:conflict",
                        "urn:dart:problem:domain",
                        "urn:dart:problem:internal",
                        "urn:dart:problem:method-not-allowed",
                        "urn:dart:problem:not-found",
                        "urn:dart:problem:upstream",
                        "urn:dart:problem:validation"
                      ],
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "The service is not configured."
          }
        },
        "security": [
          {
            "GatewayBearer": []
          }
        ],
        "summary": "Query Dataset"
      }
    },
    "/v0/metadata/contact": {
      "post": {
        "description": "Accept a JSON contact lookup for dashboard form clients.",
        "operationId": "post_contact_metadata_v0_metadata_contact_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactLookup"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactMetadata"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "detail": {
                      "maxLength": 2000,
                      "minLength": 1,
                      "type": "string"
                    },
                    "instance": {
                      "maxLength": 2000,
                      "type": "string"
                    },
                    "status": {
                      "maximum": 599.0,
                      "minimum": 400.0,
                      "type": "integer"
                    },
                    "title": {
                      "maxLength": 200,
                      "minLength": 1,
                      "type": "string"
                    },
                    "type": {
                      "enum": [
                        "urn:dart:problem:authentication",
                        "urn:dart:problem:configuration",
                        "urn:dart:problem:conflict",
                        "urn:dart:problem:domain",
                        "urn:dart:problem:internal",
                        "urn:dart:problem:method-not-allowed",
                        "urn:dart:problem:not-found",
                        "urn:dart:problem:upstream",
                        "urn:dart:problem:validation"
                      ],
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Bearer authentication failed."
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "detail": {
                      "maxLength": 2000,
                      "minLength": 1,
                      "type": "string"
                    },
                    "instance": {
                      "maxLength": 2000,
                      "type": "string"
                    },
                    "status": {
                      "maximum": 599.0,
                      "minimum": 400.0,
                      "type": "integer"
                    },
                    "title": {
                      "maxLength": 200,
                      "minLength": 1,
                      "type": "string"
                    },
                    "type": {
                      "enum": [
                        "urn:dart:problem:authentication",
                        "urn:dart:problem:configuration",
                        "urn:dart:problem:conflict",
                        "urn:dart:problem:domain",
                        "urn:dart:problem:internal",
                        "urn:dart:problem:method-not-allowed",
                        "urn:dart:problem:not-found",
                        "urn:dart:problem:upstream",
                        "urn:dart:problem:validation"
                      ],
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "The request is syntactically or semantically invalid."
          },
          "502": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "detail": {
                      "maxLength": 2000,
                      "minLength": 1,
                      "type": "string"
                    },
                    "instance": {
                      "maxLength": 2000,
                      "type": "string"
                    },
                    "status": {
                      "maximum": 599.0,
                      "minimum": 400.0,
                      "type": "integer"
                    },
                    "title": {
                      "maxLength": 200,
                      "minLength": 1,
                      "type": "string"
                    },
                    "type": {
                      "enum": [
                        "urn:dart:problem:authentication",
                        "urn:dart:problem:configuration",
                        "urn:dart:problem:conflict",
                        "urn:dart:problem:domain",
                        "urn:dart:problem:internal",
                        "urn:dart:problem:method-not-allowed",
                        "urn:dart:problem:not-found",
                        "urn:dart:problem:upstream",
                        "urn:dart:problem:validation"
                      ],
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "An upstream dependency failed."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "detail": {
                      "maxLength": 2000,
                      "minLength": 1,
                      "type": "string"
                    },
                    "instance": {
                      "maxLength": 2000,
                      "type": "string"
                    },
                    "status": {
                      "maximum": 599.0,
                      "minimum": 400.0,
                      "type": "integer"
                    },
                    "title": {
                      "maxLength": 200,
                      "minLength": 1,
                      "type": "string"
                    },
                    "type": {
                      "enum": [
                        "urn:dart:problem:authentication",
                        "urn:dart:problem:configuration",
                        "urn:dart:problem:conflict",
                        "urn:dart:problem:domain",
                        "urn:dart:problem:internal",
                        "urn:dart:problem:method-not-allowed",
                        "urn:dart:problem:not-found",
                        "urn:dart:problem:upstream",
                        "urn:dart:problem:validation"
                      ],
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "The service is not configured."
          }
        },
        "security": [
          {
            "GatewayBearer": []
          }
        ],
        "summary": "Post Contact Metadata"
      }
    },
...
  }
}